+++ src/ThumbCreator.php
@@ -19,6 +19,7 @@ use Intervention\Image\Constraint;
 use Intervention\Image\Exception\NotReadableException;
 use Intervention\Image\Image;
 use Intervention\Image\ImageManager;
+use InvalidArgumentException;
 use Symfony\Component\Filesystem\Exception\IOException;
 use Thumber\Exception\NotReadableImageException;
 use Thumber\Exception\UnsupportedImageTypeException;
@@ -41,7 +42,6 @@ class ThumbCreator
     protected $Filesystem;
 
     /**
-     * `ImageManager` instance
      * @var \Intervention\Image\ImageManager
      */
     public $ImageManager;
@@ -58,7 +58,7 @@ class ThumbCreator
     /**
      * Callbacks that will be called by the `save()` method to create the
      * thumbnail
-     * @var array
+     * @var array<callable>
      */
     protected $callbacks = [];
 
@@ -81,9 +81,6 @@ class ThumbCreator
      *  thumbnail. It can be a full path or a remote url
      * @throws \Tools\Exception\FileNotExistsException
      * @throws \Tools\Exception\NotReadableException
-     * @uses $ImageManager
-     * @uses $arguments
-     * @uses $path
      */
     public function __construct(string $path)
     {
@@ -95,10 +92,9 @@ class ThumbCreator
 
     /**
      * Internal method to get default options for the `save()` method
-     * @param array $options Passed options
+     * @param array<string, mixed> $options Passed options
      * @param string $path Path to use
-     * @return array Passed options with default options
-     * @uses $path
+     * @return array<string, mixed> Passed options with default options
      */
     protected function getDefaultSaveOptions(array $options = [], string $path = ''): array
     {
@@ -119,8 +115,6 @@ class ThumbCreator
      * @return \Intervention\Image\Image
      * @throws \Thumber\Exception\NotReadableImageException
      * @throws \Thumber\Exception\UnsupportedImageTypeException
-     * @uses $ImageManager
-     * @uses $path
      */
     protected function getImageInstance(): Image
     {
@@ -144,25 +138,23 @@ class ThumbCreator
      *
      * You can define optional coordinates to move the top-left corner of the
      *  cutout to a certain position.
-     * @param int|null $width Required width
-     * @param int|null $heigth Required heigth
-     * @param array $options Options for the thumbnail
+     * @param int $width Required width
+     * @param int $heigth Required heigth. If not specified, the width will be used
+     * @param array<string, int> $options Options for the thumbnail
      * @return self
      * @see https://github.com/mirko-pagliai/php-thumber/wiki/How-to-use-ThumbCreator-and-create-thumbnails#crop
-     * @uses $arguments
-     * @uses $callbacks
      */
-    public function crop(?int $width = null, ?int $heigth = null, array $options = []): ThumbCreator
+    public function crop(int $width, int $heigth = 0, array $options = []): ThumbCreator
     {
         $heigth = $heigth ?: $width;
-        $width = $width ?: $heigth;
-        $options += ['x' => null, 'y' => null];
+        $options += ['x' => 0, 'y' => 0];
+        Exceptionist::isPositive($width, sprintf('You have to set at least the width for the `%s()` method', __METHOD__), InvalidArgumentException::class);
 
-        //Adds arguments
         $this->arguments[] = [__FUNCTION__, $width, $heigth, $options];
-
-        //Adds the callback
         $this->callbacks[] = function (Image $imageInstance) use ($width, $heigth, $options): Image {
+            Exceptionist::isInt($options['x'], 'The `x` option must be an integer', InvalidArgumentException::class);
+            Exceptionist::isInt($options['y'], 'The `y` option must be an integer', InvalidArgumentException::class);
+
             return $imageInstance->crop($width, $heigth, $options['x'], $options['y']);
         };
 
@@ -173,26 +165,21 @@ class ThumbCreator
      * Resizes the image, combining cropping and resizing to format image in a
      *  smart way. It will find the best fitting aspect ratio on the current
      *  image automatically, cut it out and resize it to the given dimension
-     * @param int|null $width Required width
-     * @param int|null $heigth Required heigth
-     * @param array $options Options for the thumbnail
+     * @param int $width Required width
+     * @param int $heigth Required heigth. If not specified, the width will be used
+     * @param array<string, mixed> $options Options for the thumbnail
      * @return self
      * @see https://github.com/mirko-pagliai/php-thumber/wiki/How-to-use-ThumbCreator-and-create-thumbnails#fit
-     * @uses $arguments
-     * @uses $callbacks
      */
-    public function fit(?int $width = null, ?int $heigth = null, array $options = [])
+    public function fit(int $width, int $heigth = 0, array $options = [])
     {
         $heigth = $heigth ?: $width;
-        $width = $width ?: $heigth;
         $options += ['position' => 'center', 'upsize' => true];
+        Exceptionist::isPositive($width, sprintf('You have to set at least the width for the `%s()` method', __METHOD__), InvalidArgumentException::class);
 
-        //Adds arguments
         $this->arguments[] = [__FUNCTION__, $width, $heigth, $options];
-
-        //Adds the callback
         $this->callbacks[] = function (Image $imageInstance) use ($width, $heigth, $options): Image {
-            return $imageInstance->fit($width, $heigth, function (Constraint $constraint) use ($options) {
+            return $imageInstance->fit($width, $heigth, function (Constraint $constraint) use ($options): void {
                 if ($options['upsize']) {
                     $constraint->upsize();
                 }
@@ -204,24 +191,21 @@ class ThumbCreator
 
     /**
      * Resizes the image
-     * @param int|null $width Required width
-     * @param int|null $heigth Required heigth
-     * @param array $options Options for the thumbnail
+     * @param int $width Required width
+     * @param int $heigth Required heigth. If not specified, the width will be used
+     * @param array<string, bool> $options Options for the thumbnail
      * @return self
      * @see https://github.com/mirko-pagliai/php-thumber/wiki/How-to-use-ThumbCreator-and-create-thumbnails#resize
-     * @uses $arguments
-     * @uses $callbacks
      */
-    public function resize(?int $width = null, ?int $heigth = null, array $options = [])
+    public function resize(int $width, int $heigth = 0, array $options = [])
     {
+        $heigth = $heigth ?: $width;
         $options += ['aspectRatio' => true, 'upsize' => true];
+        Exceptionist::isPositive($width, sprintf('You have to set at least the width for the `%s()` method', __METHOD__), InvalidArgumentException::class);
 
-        //Adds arguments
         $this->arguments[] = [__FUNCTION__, $width, $heigth, $options];
-
-        //Adds the callback
         $this->callbacks[] = function (Image $imageInstance) use ($width, $heigth, $options): Image {
-            return $imageInstance->resize($width, $heigth, function (Constraint $constraint) use ($options) {
+            return $imageInstance->resize($width, $heigth, function (Constraint $constraint) use ($options): void {
                 if ($options['aspectRatio']) {
                     $constraint->aspectRatio();
                 }
@@ -240,22 +224,19 @@ class ThumbCreator
      *  resizing is going to happen. Set the mode to relative to add or subtract
      *  the given width or height to the actual image dimensions. You can also
      *  pass a background color for the emerging area of the image
-     * @param int|null $width Required width
-     * @param int|null $heigth Required heigth
-     * @param array $options Options for the thumbnail
+     * @param int $width Required width
+     * @param int $heigth Required heigth. If not specified, the width will be used
+     * @param array<string, mixed> $options Options for the thumbnail
      * @return self
      * @see https://github.com/mirko-pagliai/php-thumber/wiki/How-to-use-ThumbCreator-and-create-thumbnails#resizecanvas
-     * @uses $arguments
-     * @uses $callbacks
      */
-    public function resizeCanvas(?int $width, ?int $heigth = null, array $options = [])
+    public function resizeCanvas(int $width, int $heigth = 0, array $options = [])
     {
+        $heigth = $heigth ?: $width;
         $options += ['anchor' => 'center', 'relative' => false, 'bgcolor' => '#ffffff'];
+        Exceptionist::isPositive($width, sprintf('You have to set at least the width for the `%s()` method', __METHOD__), InvalidArgumentException::class);
 
-        //Adds arguments
         $this->arguments[] = [__FUNCTION__, $width, $heigth, $options];
-
-        //Adds the callback
         $this->callbacks[] = function (Image $imageInstance) use ($width, $heigth, $options): Image {
             return $imageInstance->resizeCanvas($width, $heigth, $options['anchor'], $options['relative'], $options['bgcolor']);
         };
@@ -265,17 +246,11 @@ class ThumbCreator
 
     /**
      * Saves the thumbnail and returns its path
-     * @param array $options Options for saving
+     * @param array<string, mixed> $options Options for saving
      * @return string Thumbnail path
      * @see https://github.com/mirko-pagliai/php-thumber/wiki/How-to-use-ThumbCreator-and-create-thumbnails#save-the-thumbnail
      * @throws \BadMethodCallException
      * @throws \Tools\Exception\NotWritableException
-     * @uses getDefaultSaveOptions()
-     * @uses getImageInstance()
-     * @uses $arguments
-     * @uses $callbacks
-     * @uses $path
-     * @uses $target
      */
     public function save(array $options = []): string
     {