+++ tests/TestCase/Exception/NotReadableImageExceptionTest.php
@@ -26,7 +26,7 @@ class NotReadableImageExceptionTest extends TestCase
      * Test for the exception
      * @ŧest
      */
-    public function testException()
+    public function testException(): void
     {
         try {
             throw new NotReadableImageException();